+++ extras/test/ReadBufferingClientTest.cpp
@@ -38,7 +38,8 @@ TEST_CASE("ReadBufferingClient") {
 
         CHECK(client.available() == 0);
         CHECK(log.str() ==
-              "read(4) -> 0 [timeout]"
+              "available() -> 0"
+              "read() -> -1"
               "available() -> 0");
       }
 
@@ -52,6 +53,7 @@ TEST_CASE("ReadBufferingClient") {
 
         CHECK(client.available() == 7);
         CHECK(log.str() ==
+              "available() -> 8"
               "read(4) -> 4"
               "available() -> 4");
       }
@@ -83,7 +85,9 @@ TEST_CASE("ReadBufferingClient") {
         int result = client.peek();
 
         CHECK(result == 'B');
-        CHECK(log.str() == "read(4) -> 2 [timeout]");
+        CHECK(log.str() ==
+              "available() -> 2"
+              "read(2) -> 2");
       }
     }
 
@@ -98,8 +102,10 @@ TEST_CASE("ReadBufferingClient") {
 
         CHECK(result == "ABCDEFG");
         CHECK(log.str() ==
+              "available() -> 7"
               "read(4) -> 4"
-              "read(4) -> 3 [timeout]");
+              "available() -> 3"
+              "read(3) -> 3");
       }
 
       SUBCASE("returns -1 when empty") {
@@ -108,7 +114,9 @@ TEST_CASE("ReadBufferingClient") {
         int result = client.read();
 
         CHECK(result == -1);
-        CHECK(log.str() == "read(4) -> 0 [timeout]");
+        CHECK(log.str() ==
+              "available() -> 0"
+              "read() -> -1");
       }
     }
 
@@ -120,7 +128,16 @@ TEST_CASE("ReadBufferingClient") {
         size_t result = client.readBytes(&c, 1);
 
         CHECK(result == 0);
-        CHECK(log.str() == "read(4) -> 0 [timeout]");
+
+#if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
+        CHECK(log.str() ==
+              "available() -> 0"
+              "read(1) -> 0 [timeout]");
+#else
+        CHECK(log.str() ==
+              "available() -> 0"
+              "read() -> -1");  // [timeout] from timedRead()
+#endif
       }
 
       SUBCASE("reads 4 bytes when requested one") {
@@ -131,7 +148,9 @@ TEST_CASE("ReadBufferingClient") {
 
         CHECK(c == 'A');
         CHECK(result == 1);
-        CHECK(log.str() == "read(4) -> 4");
+        CHECK(log.str() ==
+              "available() -> 7"
+              "read(4) -> 4");
       }
 
       SUBCASE("copy one byte from buffer") {
@@ -143,7 +162,9 @@ TEST_CASE("ReadBufferingClient") {
 
         CHECK(c == 'B');
         CHECK(result == 1);
-        CHECK(log.str() == "read(4) -> 4");
+        CHECK(log.str() ==
+              "available() -> 8"
+              "read(4) -> 4");
       }
 
       SUBCASE("copy content from buffer then bypass buffer") {
@@ -156,7 +177,9 @@ TEST_CASE("ReadBufferingClient") {
         CHECK(c == std::string("BCDEFGH"));
         CHECK(result == 7);
         CHECK(log.str() ==
+              "available() -> 8"
               "read(4) -> 4"
+              "available() -> 4"
               "read(4) -> 4");
       }
 
@@ -170,7 +193,9 @@ TEST_CASE("ReadBufferingClient") {
         CHECK(c == std::string("BCDE"));
         CHECK(result == 4);
         CHECK(log.str() ==
+              "available() -> 8"
               "read(4) -> 4"
+              "available() -> 4"
               "read(4) -> 4");
       }
 
@@ -182,9 +207,20 @@ TEST_CASE("ReadBufferingClient") {
         size_t result = client.readBytes(&c, 1);
 
         CHECK(result == 0);
+
+#if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
         CHECK(log.str() ==
-              "read(4) -> 1 [timeout]"
-              "read(4) -> 0 [timeout]");
+              "available() -> 1"
+              "read(1) -> 1"
+              "available() -> 0"
+              "read(1) -> 0 [timeout]");
+#else
+        CHECK(log.str() ==
+              "available() -> 1"
+              "read() -> 65"
+              "available() -> 0"
+              "read() -> -1");  // [timeout] from timedRead()
+#endif
       }
     }
 
@@ -202,7 +238,9 @@ TEST_CASE("ReadBufferingClient") {
       int result = dup.read();
 
       CHECK(result == 'B');
-      CHECK(log.str() == "read(4) -> 4");
+      CHECK(log.str() ==
+            "available() -> 8"
+            "read(4) -> 4");
     }
   }
 
@@ -264,6 +302,8 @@ TEST_CASE("ReadBufferingClient") {
     CHECK(client.readBytes(&c[3], 1) == 1);
 
     CHECK(c == std::string("{\"heEFGH"));
-    CHECK(log.str() == "read(64) -> 28 [timeout]");
+    CHECK(log.str() ==
+          "available() -> 28"
+          "read(28) -> 28");
   }
 }