+++ extras/test/ReadBufferingStreamTest.cpp
@@ -33,11 +33,10 @@ TEST_CASE("ReadBufferingStream") {
 
         CHECK(n == -1);
         CHECK(stream.available() == 0);
-#if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
         CHECK(log.str() ==
-              "readBytes(4) -> 0 [timeout]"
+              "available() -> 0"
+              "read() -> -1"
               "available() -> 0");
-#endif
       }
 
       SUBCASE("same a upstream") {
@@ -56,6 +55,7 @@ TEST_CASE("ReadBufferingStream") {
         CHECK(stream.available() == 7);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
         CHECK(log.str() ==
+              "available() -> 8"
               "readBytes(4) -> 4"
               "available() -> 4");
 #endif
@@ -89,7 +89,9 @@ TEST_CASE("ReadBufferingStream") {
 
         CHECK(result == 'B');
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-        CHECK(log.str() == "readBytes(4) -> 2 [timeout]");
+        CHECK(log.str() ==
+              "available() -> 2"
+              "readBytes(2) -> 2");
 #endif
       }
     }
@@ -106,8 +108,10 @@ TEST_CASE("ReadBufferingStream") {
         CHECK(result == "ABCDEFG");
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
         CHECK(log.str() ==
+              "available() -> 7"
               "readBytes(4) -> 4"
-              "readBytes(4) -> 3 [timeout]");
+              "available() -> 3"
+              "readBytes(3) -> 3");
 #endif
       }
 
@@ -118,7 +122,9 @@ TEST_CASE("ReadBufferingStream") {
 
         CHECK(result == -1);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-        CHECK(log.str() == "readBytes(4) -> 0 [timeout]");
+        CHECK(log.str() ==
+              "available() -> 0"
+              "read() -> -1");
 #endif
       }
     }
@@ -132,7 +138,9 @@ TEST_CASE("ReadBufferingStream") {
 
         CHECK(result == 0);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-        CHECK(log.str() == "readBytes(4) -> 0 [timeout]");
+        CHECK(log.str() ==
+              "available() -> 0"
+              "readBytes(1) -> 0 [timeout]");
 #endif
       }
 
@@ -145,7 +153,9 @@ TEST_CASE("ReadBufferingStream") {
         CHECK(c == 'A');
         CHECK(result == 1);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-        CHECK(log.str() == "readBytes(4) -> 4");
+        CHECK(log.str() ==
+              "available() -> 7"
+              "readBytes(4) -> 4");
 #endif
       }
 
@@ -159,7 +169,9 @@ TEST_CASE("ReadBufferingStream") {
         CHECK(c == 'B');
         CHECK(result == 1);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-        CHECK(log.str() == "readBytes(4) -> 4");
+        CHECK(log.str() ==
+              "available() -> 8"
+              "readBytes(4) -> 4");
 #endif
       }
 
@@ -174,7 +186,9 @@ TEST_CASE("ReadBufferingStream") {
         CHECK(result == 7);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
         CHECK(log.str() ==
+              "available() -> 8"
               "readBytes(4) -> 4"
+              "available() -> 4"
               "readBytes(4) -> 4");
 #endif
       }
@@ -190,7 +204,9 @@ TEST_CASE("ReadBufferingStream") {
         CHECK(result == 4);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
         CHECK(log.str() ==
+              "available() -> 8"
               "readBytes(4) -> 4"
+              "available() -> 4"
               "readBytes(4) -> 4");
 #endif
       }
@@ -205,8 +221,10 @@ TEST_CASE("ReadBufferingStream") {
         CHECK(result == 0);
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
         CHECK(log.str() ==
-              "readBytes(4) -> 1 [timeout]"
-              "readBytes(4) -> 0 [timeout]");
+              "available() -> 1"
+              "readBytes(1) -> 1"
+              "available() -> 0"
+              "readBytes(1) -> 0 [timeout]");
 #endif
       }
     }
@@ -226,7 +244,9 @@ TEST_CASE("ReadBufferingStream") {
 
       CHECK(result == 'B');
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-      CHECK(log.str() == "readBytes(4) -> 4");
+      CHECK(log.str() ==
+            "available() -> 8"
+            "readBytes(4) -> 4");
 #endif
     }
   }
@@ -290,7 +310,9 @@ TEST_CASE("ReadBufferingStream") {
 
     CHECK(c == std::string("{\"heEFGH"));
 #if STREAMUTILS_STREAM_READBYTES_IS_VIRTUAL
-    CHECK(log.str() == "readBytes(64) -> 28 [timeout]");
+    CHECK(log.str() ==
+          "available() -> 28"
+          "readBytes(28) -> 28");
 #endif
   }
 }