+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 StreamUtils - Change log
 ========================
 
+HEAD
+----
+
+* Fixed unwanted waits in `ReadBufferingClient` and `ReadBufferingStream`.
+
 1.4.0 (2020/03/30)
 -----